schema.sql: Initial SQL schema.
This SQL file creates the tables for clients, budgets, transactions, categories,
and stocks. According to our requirements, we want the user to be able to create
their own categories in the future. However, this adds complexity to the db, so
this feature will be added after initial testing. Furthermore, stock history is
not tracked at the moment as this is out of the scope for core functionality.