init-db: POSIX shell script to manage PostgreSQL database.
This is a script for Unix-based systems. It creates a .db file in the current
directory to store the database. This script allows the user to start, stop,
clean, and init the database. There is a hack in the script that makes it so all
users on the system can access the `/run/postgresql` file which is normally
owned by the `postgres` user on the system. This hack SHOULD NOT be used in
production. Instead the user should take proper security precautions and run the
database as `postgres`.