stock.go: Removed json parser in favor of RegEx.
Got rid of the json parser since it was unnecessary bloat imo. We are only using
a minimal subset of the response and the response is awful, so its best to use
that. Moved json response fetcher to another function and added the RegEx for
the current price and previous close.